+++ tests/kind/received_channel_metadata.rs
@@ -6,7 +6,8 @@ use nostrtalk::{
 use url::Url;
 
 use super::*;
-use crate::helpers::spawn_app;
+use crate::common::{make_channel_creation_event, make_channel_metadata_event};
+use crate::spawn_app;
 
 #[tokio::test]
 async fn channel_metadata_ok() {