+++ tests/kind/received_channel_msg.rs
@@ -1,13 +1,15 @@
+use futures_util::StreamExt;
+use nostr::Keys;
 use nostrtalk::{
-    db::MessageStatus,
-    net::handle_event,
-    types::{ChatMessage, UserMessage},
+    db::{ChannelCache, DbChannelMessage, MessageStatus},
+    net::{handle_event, BackendEvent},
+    types::{ChatMessage, SubName, UserMessage},
 };
 use url::Url;
 
 use super::dm_helpers::*;
-use super::*;
-use crate::helpers::spawn_app;
+use crate::common::{make_channel_msg_event, random_contact_channel_msg_event};
+use crate::spawn_app;
 
 /// Tests for Received event of Kind::ChannelMessage
 
@@ -17,11 +19,11 @@ async fn channel_msg_user_to_channel() {
     let mut test_app = spawn_app().await;
     let (mut output, mut rx) = futures::channel::mpsc::channel(5);
     let url = Url::parse("ws://192.168.15.15:8080").unwrap();
+
+    let cache = test_app.insert_random_channel_cache().await;
+    let channel_id = cache.channel_id;
+
     let msg_content: String = "hey group!".into();
-    let channel_id = nostr::EventId::from_hex(
-        "8233a5d8e27a9415d22c974d70935011664ada55ae3152bd10d697d3a3c74f67",
-    )
-    .unwrap();
     let ns_event = make_channel_msg_event(&test_app.keys, &channel_id, Some(&url), &msg_content);
     let event_hash = ns_event.id.clone();
     let subscription_id = nostr::SubscriptionId::new("testing");
@@ -42,6 +44,14 @@ async fn channel_msg_user_to_channel() {
 
     let db_event =
         assert_channel_msg_in_database(&test_app, &event_hash, &channel_id, 1, &msg_content).await;
+
+    if let Some(event) = rx.next().await {
+        if let BackendEvent::ChannelCacheUpdated(_) = event {
+            // Ok
+        } else {
+            panic!("Unexpected event: {:?}", event);
+        }
+    }
 
     if let Some(event) = rx.next().await {
         if let BackendEvent::ReceivedChannelMessage(rcv_channel_id, chat_message) = &event {
@@ -79,11 +89,11 @@ async fn channel_msg_anyone_to_channel() {
     let (mut output, mut rx) = futures::channel::mpsc::channel(5);
     let url = Url::parse("ws://192.168.15.15:8080").unwrap();
     let sender_keys = Keys::generate();
+
+    let cache = test_app.insert_random_channel_cache().await;
+    let channel_id = cache.channel_id;
+
     let msg_content: String = "hey group".into();
-    let channel_id = nostr::EventId::from_hex(
-        "8233a5d8e27a9415d22c974d70935011664ada55ae3152bd10d697d3a3c74f67",
-    )
-    .unwrap();
     let ns_event = make_channel_msg_event(&sender_keys, &channel_id, Some(&url), &msg_content);
     let event_hash = ns_event.id.clone();
     let subscription_id = nostr::SubscriptionId::new("testing");
@@ -106,6 +116,14 @@ async fn channel_msg_anyone_to_channel() {
         assert_channel_msg_in_database(&test_app, &event_hash, &channel_id, 1, &msg_content).await;
 
     if let Some(event) = rx.next().await {
+        if let BackendEvent::ChannelCacheUpdated(_) = event {
+            // Ok
+        } else {
+            panic!("Unexpected event: {:?}", event);
+        }
+    }
+
+    if let Some(event) = rx.next().await {
         if let BackendEvent::ReceivedChannelMessage(rcv_channel_id, chat_message) = &event {
             assert_eq!(rcv_channel_id, &channel_id);
 
@@ -126,6 +144,66 @@ async fn channel_msg_anyone_to_channel() {
             }
         } else {
             panic!("Unexpected event: {:?}", event);
+        }
+    }
+}
+
+#[tokio::test]
+async fn search_channel_msg() {
+    // PREPARE
+    let mut test_app = spawn_app().await;
+    let (mut output, mut rx) = futures::channel::mpsc::channel(20);
+    let url = Url::parse("ws://192.168.15.15:8080").unwrap();
+    let cache = test_app.insert_random_channel_cache().await;
+    let channel_id = cache.channel_id;
+
+    let subscription_id =
+        nostr::SubscriptionId::new(SubName::src_channel_details(&channel_id).to_string());
+    let messages = vec!["hey group", "hey john", "yo", "lets go", "everyone"];
+
+    // PERFORM
+    for msg in &messages {
+        let result = handle_event(
+            &mut output,
+            &test_app.keys,
+            &mut test_app.backend,
+            url.clone(),
+            subscription_id.clone(),
+            random_contact_channel_msg_event(&channel_id, msg),
+        )
+        .await;
+        assert!(result.is_ok(), "Error handling event: {:?}", result.err());
+    }
+
+    // ASSERT
+    let db_msgs = DbChannelMessage::fetch(test_app.pool(), &channel_id)
+        .await
+        .unwrap();
+    assert_eq!(db_msgs.len(), messages.len());
+
+    let cache = ChannelCache::fetch_by_channel_id(test_app.cache_pool(), &channel_id)
+        .await
+        .unwrap()
+        .unwrap();
+    assert_eq!(cache.members.len(), messages.len());
+
+    for (idx, msg) in messages.iter().enumerate() {
+        if let Some(event) = rx.next().await {
+            if let BackendEvent::ChannelCacheUpdated(cache) = &event {
+                assert_eq!(cache.members.len(), idx + 1);
+                assert_eq!(&cache.channel_id, &channel_id);
+            } else {
+                panic!("Unexpected event: {:?}", event);
+            }
+        }
+
+        if let Some(event) = rx.next().await {
+            if let BackendEvent::ReceivedChannelMessage(rcv_channel_id, chat_message) = &event {
+                assert_eq!(rcv_channel_id, &channel_id);
+                assert_eq!(&chat_message.content(), msg);
+            } else {
+                panic!("Unexpected event: {:?}", event);
+            }
         }
     }
 }